Add BoundsMap component which sets map view by bounds
The `Map` component only supports setting and updating the map view by
changing the map `center` and `zoom` props. It is often useful to set
the map view by specifying some bounds that must be contained instead.

For example: loading a full-screen map to show a country.

This commit adds a `BoundsMap` component which extends `Map` and simply
takes a `bounds` prop (of type `src/lib/types/bounds`) instead of
`center` and `zoom` props.

Tests for this component were adapted from those for `Map`.