Lint 'object destructuring filter' consistently
The pattern ```const { map, url, ...props } = this.props``` usually
failed during linting, because the ```map``` variable is defined but
not used.

After adding the ```"varsIgnorePattern": "^_"``` option to the
```no-unused-vars``` rule, unused properies can be marked as:
```const { map: _, url, ...props } = this.props```.

I also updated the eslint and babel-eslint dependencies, because of
some relevant bugfixes.