Fix duplicate leaflet instantiation in React 18 strict mode
React 18 in strict mode simulates an extra unmount + mount cycle in every component's first mount, which results in the rendering effect to run twice with the same conditions, causing a double call to `new LeafletMap()` which leads to an exception in the leaflet library. There are multiple possibilities described in the official react communication (https://github.com/reactwg/react-18/discussions/18) to prevent it, and I tried tearing down the instance (`return () => instance.remove()`), but it led to other problems, and the code already contains teardown logic. So I went with a state-ref useRef way, where the ref keeps its value over unmounts (as opposed to useState), also described here: https://github.com/reactwg/react-18/discussions/18#discussioncomment-801681, so we can effectively prevent the double instantiation.

We still need the `const [map, setMap] = useState...` part, in order for the useEffect hook to be able trigger a rerender after all the refs are having a value.