Transpile to es2019 to increase backwards compatibility
Transpile to es2019 for more backwards compatibility. With es2020, the library preserves nullish coalescing operators in the output. Currently, it's necessary to run CRA 5 to support that, which is not compatible with React 18. For the time being it would be best if this library could work without CRA 5, out of the box.